*: move row count cache from package executor to statistics
