+++ statistics/handle/historical_stats_handler.go
@@ -75,6 +75,7 @@ func recordHistoricalStatsMeta(sctx sessionctx.Context, tableID int64, version u
 	if _, err := exec.ExecuteInternal(ctx, sql, tableID, modifyCount, count, version, source); err != nil {
 		return errors.Trace(err)
 	}
+	TableRowStatsCache.Invalidate(tableID)
 	return nil
 }
